requirements.txt
torch
torchvision
transformers
gradio
Pillow
opencv-python
scenedetect[opencv]
